extract handleTaskCompletion
